requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
pypdf==5.1.0
chromadb==0.5.23
sentence-transformers==3.2.1
numpy==2.1.3
openai==1.58.1
